added exercises

+++ semana_6/ejercicio12/src/main/java/ar/edu/info/unlp/ejercicio12/Esfera.java
@@ -18,7 +18,7 @@ public class Esfera extends Pieza {
 
     @Override
     public double getVolumen() {
-        return (double) 4 /3 * Math.PI * Math.pow(this.radio,3);
+        return 4.0 / 3.0 * Math.PI * Math.pow(this.radio,3);
     }
 
     @Override
